+++ src/asciidocEngine.ts
@@ -9,12 +9,8 @@ import { Asciidoctor } from '@asciidoctor/core'
 import { SkinnyTextDocument } from './util/document'
 import { AsciidoctorExtensionsSecurityPolicyArbiter } from './security'
 
-const FrontMatterRegex = /^---\s*[^]*?(-{3}|\.{3})\s*/
-
 export class AsciidocEngine {
   private ad?: AsciidocParser
-
-  private firstLine?: number
 
   public constructor (
     readonly extensionPreviewResourceProvider: AsciidocContributions,
@@ -35,43 +31,25 @@ export class AsciidocEngine {
     return this.ad
   }
 
-  private stripFrontmatter (text: string): { text: string, offset: number } {
-    let offset = 0
-    const frontMatterMatch = FrontMatterRegex.exec(text)
-    if (frontMatterMatch) {
-      const frontMatter = frontMatterMatch[0]
-      offset = frontMatter.split(/\r\n|\n|\r/g).length - 1
-      text = text.substr(frontMatter.length)
-    }
-    return { text, offset }
-  }
-
   public async convert (
     documentUri: vscode.Uri,
-    stripFrontmatter: boolean,
     text: string,
     context: vscode.ExtensionContext,
     editor: vscode.WebviewPanel
   ): Promise<{output: string, document?: Asciidoctor.Document}> {
     const parser = this.getEngine()
-
-    let offset = 0
-    if (stripFrontmatter) {
-      const asciidocContent = this.stripFrontmatter(text)
-      offset = asciidocContent.offset
-      text = asciidocContent.text
-    }
-
-    this.firstLine = offset
     const textDocument = await vscode.workspace.openTextDocument(documentUri)
     const { html: output, document } = await parser.convertUsingJavascript(text, textDocument, context, editor)
     return { output, document }
   }
 
-  public async export (textDocument: vscode.TextDocument, backend: AsciidoctorBuiltInBackends): Promise<{ output: string, document: Asciidoctor.Document }> {
+  public async export (
+    textDocument: vscode.TextDocument,
+    backend: AsciidoctorBuiltInBackends,
+    asciidoctorAttributes = {}
+  ): Promise<{ output: string, document: Asciidoctor.Document }> {
     const parser = this.getEngine()
-
-    return parser.export(textDocument.getText(), textDocument, backend)
+    return parser.export(textDocument.getText(), textDocument, backend, asciidoctorAttributes)
   }
 
   public load (textDocument: SkinnyTextDocument): Asciidoctor.Document {