+++ src/commands/exportAsPDF.ts
@@ -2,141 +2,279 @@ import * as vscode from 'vscode'
 import * as os from 'os'
 import * as fs from 'fs'
 import * as path from 'path'
-import { exec, spawn } from 'child_process'
+import { exec, spawn, SpawnOptions } from 'child_process'
 import { uuidv4 } from 'uuid'
-import * as zlib from 'zlib'
 import { AsciidocEngine } from '../asciidocEngine'
 import { Command } from '../commandManager'
 import { Logger } from '../logger'
 import { Asciidoctor } from '@asciidoctor/core'
-import url = require('url');
 
 export class ExportAsPDF implements Command {
   public readonly id = 'asciidoc.exportAsPDF'
+  private readonly exportAsPdfStatusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Right, 100)
 
-  constructor (private readonly engine: AsciidocEngine, private readonly logger: Logger) {
-    this.engine = engine
-    this.logger = logger
+  constructor (private readonly engine: AsciidocEngine, private readonly context: vscode.ExtensionContext, private readonly logger: Logger) {
   }
 
   public async execute () {
     const editor = vscode.window.activeTextEditor
-
-    if (editor === null || editor === undefined) { return }
+    if (editor === null || editor === undefined) {
+      return
+    }
 
     const doc = editor.document
-    const sourceName = path.parse(path.resolve(doc.fileName))
-    const pdfFilename = vscode.Uri.file(path.join(sourceName.root, sourceName.dir, sourceName.name + '.pdf'))
+    const workspaceFolder = vscode.workspace.getWorkspaceFolder(doc.uri)
+    if (workspaceFolder === undefined) {
+      await vscode.window.showWarningMessage('Unable to get the workspace folder, aborting.')
+      return
+    }
+    const workspacePath = workspaceFolder.uri.fsPath
+    const docUri = path.parse(path.resolve(doc.fileName))
+    const baseDirectory = path.join(docUri.root, docUri.dir)
+    const pdfFilename = vscode.Uri.file(path.join(baseDirectory, docUri.name + '.pdf'))
 
+    const asciidocPdfConfig = vscode.workspace.getConfiguration('asciidoc.pdf')
+    const pdfOutputUri = await vscode.window.showSaveDialog({ defaultUri: pdfFilename })
+    if (!pdfOutputUri) {
+      console.log('No output directory selected to save the PDF, aborting.')
+      return
+    }
+
+    const pdfOutputPath = pdfOutputUri.fsPath
     const text = doc.getText()
-    if (vscode.workspace.getConfiguration('asciidoc', null).get('use_asciidoctorpdf')) {
-      const docPath = path.parse(path.resolve(doc.fileName))
-      let pdfPath = ''
-
-      const pdfUri = await vscode.window.showSaveDialog({ defaultUri: pdfFilename })
-      if (pdfUri) {
-        pdfPath = pdfUri.fsPath
-      } else {
-        console.error(`ERROR: invalid pdfUri "${pdfUri}"`)
+    const pdfEnfine = asciidocPdfConfig.get('engine')
+    if (pdfEnfine === 'asciidoctor-pdf') {
+      const asciidoctorPdfCommand = await this.resolveAsciidoctorPdfCommand(asciidocPdfConfig, workspacePath)
+      if (asciidoctorPdfCommand === undefined) {
         return
       }
-      const asciidoctorPDFCommand = vscode.workspace
-        .getConfiguration('asciidoc', null)
-        .get('asciidoctorpdf_command', 'asciidoctor-pdf')
+      const asciidoctorPdfCommandArgs = asciidocPdfConfig.get<string[]>('asciidoctorPdfCommandArgs', [])
+      const defaultArgs = [
+        '-q', // quiet
+        '-B',
+        `"${baseDirectory.replace('"', '\\"')}"`, // base directory
+        '-o',
+        `"${pdfOutputPath.replace('"', '\\"')}"`, // output file
+      ]
+      const args = defaultArgs.concat(asciidoctorPdfCommandArgs)
+        .concat(['-']) // read from stdin
 
-      const adocPDFCmdArray = asciidoctorPDFCommand
-        .split(/(\s+)/)
-        .filter(function (e) { return e.trim().length > 0 })
-
-      const adocPDFCmd = adocPDFCmdArray[0]
-
-      const adocPDFCmdArgs = adocPDFCmdArray.slice(1)
-      adocPDFCmdArgs.push('-q', '-B', '"' + docPath.dir.replace('"', '\\"') + '"',
-        '-o', '"' + pdfPath.replace('"', '\\"') + '"', '-')
-
-      const options = { shell: true, cwd: docPath.dir }
-
-      const asciidoctorPDF = spawn(adocPDFCmd, adocPDFCmdArgs, options)
-
-      asciidoctorPDF.stderr.on('data', (data) => {
-        let errorMessage = data.toString()
-        errorMessage += '\n'
-        errorMessage += 'command: ' + adocPDFCmd + ' ' + adocPDFCmdArgs.join(' ')
-        errorMessage += '\n'
-        errorMessage += 'If the asciidoctor-pdf binary is not in your PATH, you can set the full path.'
-        errorMessage += 'Go to `File -> Preferences -> User settings` and adjust the asciidoc.asciidoctorPDFCommand'
-        console.error(errorMessage)
-        vscode.window.showErrorMessage(errorMessage)
-      })
-
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      asciidoctorPDF.on('close', (code) => {
-        offerOpen(pdfPath)
-      })
-
-      asciidoctorPDF.stdin.write(text)
-      asciidoctorPDF.stdin.end()
-    } else {
-      const wkHTMLtoPDFPath = vscode.workspace
-        .getConfiguration('asciidoc')
-        .get('wkhtmltopdf_path', '')
-
-      const { output: html, document } = await this.engine.export(doc, 'html5')
-      const showTitlePage = (document?.isAttribute('showTitlePage') as unknown) as boolean // incorrect type definition in Asciidoctor.js
-      const titlePageLogo: string | undefined = document?.getAttribute('titlePageLogo')
-      const footerCenter: string | undefined = document?.getAttribute('footer-center')
-      const coverFilePath = showTitlePage ? createCoverFile(titlePageLogo, sourceName.dir, document) : undefined
-      const binaryPath = wkHTMLtoPDFPath || path.resolve(path.join(__dirname, `wkhtmltopdf-${process.platform}-${process.arch}${process.platform === 'win32' ? '.exe' : ''}`))
-
-      if (!fs.existsSync(binaryPath)) {
-        const label = await vscode.window.showInformationMessage('This feature requires wkhtmltopdf\ndo you want to download', 'Download')
-        if (label !== 'Download') { return }
-
-        const downloadSuccessful: boolean = await vscode.window.withProgress({
-          location: vscode.ProgressLocation.Window,
-          title: 'Downloading wkhtmltopdf',
-          // cancellable: true
-        }, async (progress) => {
-          progress.report({ message: 'Downloading wkhtmltopdf...' })
-          const downloadURL = `https://github.com/joaompinto/wkhtmltopdf/releases/download/v0.0.1/wkhtmltopdf-${process.platform}-${process.arch}${process.platform === 'win32' ? '.exe' : ''}.gz`
-          this.logger.log('Downloading ' + downloadURL)
-          try {
-            await downloadFile(downloadURL, binaryPath + '.gz', progress)
-            progress.report({ message: 'Unzipping wkhtmltopdf...' })
-            const ungzip = zlib.createGunzip()
-            const inp = fs.createReadStream(binaryPath + '.gz')
-            const out = fs.createWriteStream(binaryPath)
-            inp.pipe(ungzip).pipe(out)
-            fs.chmodSync(binaryPath, 0x755)
-            return true
-          } catch (err) {
-            console.error('Error downloading', downloadURL, ' ', err)
-            await vscode.window.showErrorMessage('Error installing wkhtmltopdf, ' + err.toString())
-            return false
-          }
-        })
-        if (!downloadSuccessful) {
-          // abort!
-          return
-        }
+      try {
+        this.exportAsPdfStatusBarItem.name = 'Export As PDF'
+        this.exportAsPdfStatusBarItem.text = '$(loading~spin) Generating a PDF using asciidoctor-pdf...'
+        this.exportAsPdfStatusBarItem.show()
+        await execute(asciidoctorPdfCommand.command, args, text, { shell: true, cwd: asciidoctorPdfCommand.cwd })
+        this.exportAsPdfStatusBarItem.text = '$(pass) PDF has been successfully generated!'
+        offerOpen(pdfOutputPath)
+      } catch (err) {
+        console.error('Unable to generate a PDF using asciidoctor-pdf: ', err)
+        await vscode.window.showErrorMessage(`Unable to generate a PDF using asciidoctor-pdf: ${err}`)
+      } finally {
+        this.exportAsPdfStatusBarItem.hide()
       }
-      const saveFileUri = await vscode.window.showSaveDialog({ defaultUri: pdfFilename })
-      if (saveFileUri) {
-        try {
-          const generatedPdf = await html2pdf(html, binaryPath, coverFilePath, footerCenter, saveFileUri.fsPath)
-          offerOpen(generatedPdf)
-        } catch (err) {
-          console.error('Got error', err)
-          await vscode.window.showErrorMessage('Error converting to PDF, ' + err.toString())
-        } finally {
-          if (coverFilePath) {
-            // remove temporary file
-            fs.unlinkSync(coverFilePath)
-          }
+    } else if (pdfEnfine === 'wkhtmltopdf') {
+      let wkhtmltopdfCommandPath = asciidocPdfConfig.get('wkhtmltopdfCommandPath', '')
+      if (wkhtmltopdfCommandPath === '') {
+        wkhtmltopdfCommandPath = `wkhtmltopdf${process.platform === 'win32' ? '.exe' : ''}`
+      } else {
+        /* eslint-disable-next-line no-template-curly-in-string */
+        wkhtmltopdfCommandPath = wkhtmltopdfCommandPath.replace('${workspaceFolder}', workspacePath)
+      }
+      try {
+        await commandExists(wkhtmltopdfCommandPath, { shell: true, cwd: workspacePath })
+      } catch (error) {
+        // command does not exist!
+        console.error(error)
+        const answer = await vscode.window.showInformationMessage('This feature requires wkhtmltopdf. Please download the latest version from https://wkhtmltopdf.org/downloads.html. If wkhtmltopdf is not available on your path, you can configure the path to wkhtmltopdf executable from the extension settings.', 'Download')
+        if (answer === 'Download') {
+          vscode.env.openExternal(vscode.Uri.parse('https://wkhtmltopdf.org/downloads.html'))
+        }
+        return
+      }
+      const wkhtmltopdfCommandArgs = asciidocPdfConfig.get<string[]>('wkhtmltopdfCommandArgs', [])
+      const defaultArgs = ['--enable-local-file-access', '--encoding', ' utf-8', '--javascript-delay', '1000']
+
+      const { output: html, document } = await this.engine.export(doc, 'html5', { 'data-uri@': '' })
+      const footerCenter = document?.getAttribute('footer-center')
+      if (footerCenter) {
+        defaultArgs.push('--footer-center', footerCenter)
+      }
+      const objectArgs = []
+      const showTitlePage = (document?.isAttribute('showTitlePage') as unknown) as boolean // incorrect type definition in Asciidoctor.js
+      const titlePageLogo = document?.getAttribute('titlePageLogo')
+      const coverFilePath = showTitlePage ? createCoverFile(titlePageLogo, baseDirectory, document) : undefined
+      if (coverFilePath) {
+        objectArgs.push('cover', coverFilePath)
+      }
+      // wkhtmltopdf [GLOBAL OPTION]... [OBJECT]... <output file>
+      const args = defaultArgs.concat(wkhtmltopdfCommandArgs)
+        .concat(objectArgs)
+        .concat(['-', pdfOutputPath]) // read from stdin and outputfile
+
+      try {
+        this.exportAsPdfStatusBarItem.name = 'Export As PDF'
+        this.exportAsPdfStatusBarItem.text = '$(loading~spin) Generating a PDF using wkhtmltopdf...'
+        this.exportAsPdfStatusBarItem.show()
+        await execute(wkhtmltopdfCommandPath, args, html, { shell: true, cwd: workspacePath, stdio: ['pipe', 'ignore', 'pipe'] })
+        this.exportAsPdfStatusBarItem.text = '$(pass) PDF has been successfully generated!'
+        offerOpen(pdfOutputPath)
+      } catch (err) {
+        console.error('Unable to generate a PDF using wkhtmltopdf: ', err)
+        await vscode.window.showErrorMessage(`Unable to generate a PDF using wkhtmltopdf: ${err}`)
+      } finally {
+        this.exportAsPdfStatusBarItem.hide()
+        if (coverFilePath) {
+          // remove temporary file
+          fs.unlinkSync(coverFilePath)
         }
       }
     }
   }
+
+  private async resolveAsciidoctorPdfCommand (asciidocPdfConfig, workspacePath): Promise<{ cwd: string, command: string } | undefined> {
+    let asciidoctorPdfCommandPath = asciidocPdfConfig.get('asciidoctorPdfCommandPath', '')
+    if (asciidoctorPdfCommandPath !== '') {
+      /* eslint-disable-next-line no-template-curly-in-string */
+      asciidoctorPdfCommandPath = asciidoctorPdfCommandPath.replace('${workspaceFolder}', workspacePath)
+      // use the command specified
+      return {
+        cwd: workspacePath,
+        command: asciidoctorPdfCommandPath,
+      }
+    }
+    if (this.isAsciidoctorPdfAvailable(workspacePath)) {
+      // `asciidoctor-pdf` is available
+      return {
+        cwd: workspacePath,
+        command: 'asciidoctor-pdf',
+      }
+    }
+    if (this.isBundlerAvailable()) {
+      const globalStorageUri = this.context.globalStorageUri
+      const installDirectory = path.join(globalStorageUri.fsPath, 'asciidoctor-pdf-install')
+      try {
+        await commandExists('bundle exec asciidoctor-pdf', { shell: true, cwd: installDirectory })
+        return {
+          cwd: installDirectory,
+          command: 'bundle exec asciidoctor-pdf',
+        }
+      } catch (bundleExecError) {
+        console.info(`Error while trying to execute 'bundle exec asciidoctor-pdf' from '${installDirectory}', cause: `, bundleExecError)
+        // `asciidoctor-pdf` is not available in global storage, offer to automatically install it
+        const answer = await vscode.window.showInformationMessage('This feature requires asciidoctor-pdf. Do you want to install the latest version locally using Bundler? Alternatively, you can configure the path to the asciidoctor-pdf executable from the extension settings.', 'Install locally')
+        if (answer === 'Install locally') {
+          this.exportAsPdfStatusBarItem.name = 'Asciidoctor PDF Installer'
+          this.exportAsPdfStatusBarItem.text = '$(loading~spin) Installing Asciidoctor PDF...'
+          this.exportAsPdfStatusBarItem.show()
+          try {
+            if (!fs.existsSync(installDirectory)) {
+              fs.mkdirSync(installDirectory, { recursive: true })
+            }
+            const gemfile = path.join(installDirectory, 'Gemfile')
+
+            await execute('bundle', ['config', '--local', 'path', '.bundle/gem'], undefined, { cwd: installDirectory })
+            fs.writeFileSync(gemfile, `source 'https://rubygems.org'
+
+gem 'asciidoctor-pdf'`, { encoding: 'utf8' })
+            await execute('bundle', ['install'], undefined, { cwd: installDirectory })
+            this.exportAsPdfStatusBarItem.text = '$(pass) Asciidoctor PDF installed!'
+            const answer = await vscode.window.showInformationMessage(`Successfully installed Asciidoctor PDF in ${installDirectory}`, 'Continue')
+            if (answer === 'Continue') {
+              return {
+                command: 'bundle exec asciidoctor-pdf',
+                cwd: installDirectory,
+              }
+            } else {
+              return undefined
+            }
+          } catch (err) {
+            await vscode.window.showErrorMessage(`Unable to install the latest version of asciidoctor-pdf using Bundler: ${err}`)
+            return undefined
+          } finally {
+            this.exportAsPdfStatusBarItem.hide()
+          }
+        } else {
+          return undefined
+        }
+      }
+    } else {
+      const answer = await vscode.window.showInformationMessage('This feature requires asciidoctor-pdf but the executable was not found on your PATH. Please install asciidoctor-pdf or configure the path to the executable from the extension settings.', 'Install', 'Configure')
+      if (answer === 'Configure') {
+        await vscode.commands.executeCommand('workbench.action.openSettings', '@ext:asciidoctor.asciidoctor-vscode asciidoctorPdfCommand')
+      } else if (answer === 'Install') {
+        await vscode.env.openExternal(vscode.Uri.parse('https://docs.asciidoctor.org/pdf-converter/latest/install/'))
+      }
+      return undefined
+    }
+  }
+
+  private async isBundlerAvailable (): Promise<boolean> {
+    try {
+      await commandExists('bundle', { shell: true })
+      return true
+    } catch (err) {
+      // unable to find `bundle`, Bundler is probably not installed
+      return false
+    }
+  }
+
+  private async isAsciidoctorPdfAvailable (cwd: string) {
+    try {
+      await commandExists('asciidoctor-pdf', { shell: true, cwd })
+      return true
+    } catch (err) {
+      // command does not exist
+      console.warn(err)
+      return false
+    }
+  }
+}
+
+function commandExists (command: string, options: SpawnOptions): Promise<{ stdout: string, code: number }> {
+  const childProcess = spawn(command, ['--version'], { env: process.env, ...options })
+  return new Promise(function (resolve, reject) {
+    const stdoutOutput = []
+    childProcess.stdout.on('data', (data) => {
+      stdoutOutput.push(data)
+    })
+    childProcess.on('close', function (code) {
+      if (code === 0) {
+        resolve({
+          stdout: stdoutOutput.join('\n'),
+          code,
+        })
+      } else {
+        reject(new Error(`command failed: ${command}`))
+      }
+    })
+    childProcess.on('error', function (err) {
+      reject(err)
+    })
+  })
+}
+
+function execute (command: string, args: string[], input: string | undefined, options: SpawnOptions): Promise<boolean> {
+  return new Promise(function (resolve, reject) {
+    const childProcess = spawn(command, args, { env: process.env, ...options })
+    const stderrOutput = []
+    childProcess.stderr.on('data', (data) => {
+      stderrOutput.push(data)
+    })
+    childProcess.on('close', function (code) {
+      if (code === 0) {
+        resolve(true)
+      } else {
+        reject(new Error(`command failed: ${command} ${args.join(' ')}\n${stderrOutput.join('\n')}`))
+      }
+    })
+    childProcess.on('error', function (err) {
+      reject(err)
+    })
+    if (input !== undefined) {
+      childProcess.stdin.write(input)
+      childProcess.stdin.end()
+    }
+  })
 }
 
 export function _generateCoverHtmlContent (
@@ -182,57 +320,14 @@ function createCoverFile (titlePageLogo: string, baseDir: string, document: Asci
   return tmpFilePath
 }
 
-async function downloadFile (downloadURL: string, filename: string, progress): Promise<void> {
-  // load "follow-redirects" only when needed (because this module cannot be loaded in a browser environment)
-  const followRedirects = await import('follow-redirects')
-  return new Promise((resolve, reject) => {
-    const downloadOptions = url.parse(downloadURL)
-    const wstream = fs.createWriteStream(filename)
-    let totalDownloaded = 0
-    // Proxy support needs to be reworked
-    // var proxy = process.env.http_proxy || vscode.workspace.getConfiguration("http", null)["proxy"].trim();
-    // var proxyStrictSSL = vscode.workspace.getConfiguration("http", null)["proxyStrictSSL"];
-    // if (proxy != '')
-    // {
-    //   var agent = new HttpsProxyAgent(proxy);
-    //   downloadOptions.agent = agent
-    //   downloadOptions.rejectUnauthorized = proxyStrictSSL
-    // }
-    followRedirects.https.get(downloadOptions, (resp) => {
-      const contentSize = resp.headers['content-length']
-      if (resp.statusCode !== 200) {
-        wstream.end()
-        fs.unlinkSync(filename)
-        return reject(new Error('http error' + resp.statusCode))
-      }
-
-      // A chunk of data has been recieved.
-      resp.on('data', (chunk) => {
-        totalDownloaded += chunk.length
-        progress.report({ message: 'Downloading wkhtmltopdf ... ' + ((totalDownloaded / contentSize) * 100.0).toFixed(0) + '%' })
-        wstream.write(chunk)
-      })
-
-      // The whole response has been received. Print out the result.
-      resp.on('end', () => {
-        wstream.end()
-        resolve()
-      })
-    }).on('error', (err) => {
-      console.error('Error: ' + err.message)
-      reject(err)
-    })
-  })
-}
-
 function offerOpen (destination) {
   // Saving the JSON that represents the document to a temporary JSON-file.
   vscode.window.showInformationMessage(('Successfully converted to ' + path.basename(destination)), 'Open File').then((label: string) => {
     if (label === 'Open File') {
       switch (process.platform) {
-      // Use backticks for unix systems to run the open command directly
-      // This avoids having to wrap the command AND path in quotes which
-      // breaks if there is a single quote (') in the path
+        // Use backticks for unix systems to run the open command directly
+        // This avoids having to wrap the command AND path in quotes which
+        // breaks if there is a single quote (') in the path
         case 'win32':
           exec(`"${destination.replace('"', '\\"')}"`)
           break
@@ -247,36 +342,5 @@ function offerOpen (destination) {
           break
       }
     }
-  })
-}
-
-export async function html2pdf (
-  html: string,
-  binaryPath: string,
-  coverFilePath: string | undefined,
-  footerCenter: string | undefined,
-  filename: string
-): Promise<string> {
-  const documentPath = path.dirname(filename)
-
-  return new Promise((resolve, reject) => {
-    const cmdArguments = ['--encoding', ' utf-8', '--javascript-delay', '1000']
-    if (footerCenter) {
-      cmdArguments.push('--footer-center', footerCenter)
-    }
-    if (coverFilePath) {
-      cmdArguments.push('cover', coverFilePath)
-    }
-    cmdArguments.push('-', filename)
-    const command = spawn(binaryPath, cmdArguments, { cwd: documentPath, stdio: ['pipe', 'ignore', 'pipe'] })
-    let errorData = ''
-    command.stdin.write(html)
-    command.stdin.end()
-    command.stderr.on('data', (data) => {
-      errorData += data
-    })
-    command.on('close', (code) => {
-      if (code === 0) { resolve(filename) } else { reject(errorData) }
-    })
   })
 }